Modificado html y css, separado en modulos
